pyproject.toml
[tool.poetry]
name = "djang-auth-service"
version = "0.2.0"
description = ""
authors = ["Василий Серебрянников <svaaugust@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
mypy = "^1.12.0"
commitizen = "^3.29.1"
faker = "^30.3.0"
ruff = "^0.6.9"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
